Split REST API tests
Split the tests into smaller, more handlable files
